Add Type_UnboundRecordField error.
Note: right now the hint only says "did you mean X" where X is a
potential valid name for this field (allowing the type checker to
associate this inline record declaration with a type). It'd be nice
to have the name of the type itself for which this field is valid.