Add Type_UnboundModule error.
Hint containing ocamlc/ocamlbuild/oasis support. Not perfect yet, you
can still encounter errors related to this like:
"No implementations provided for the following modules" which require a
different solution. We're also not using the suggestion right now
because we can't trigger a Hint from the compiler (even a slight typo
with the linked library doesn't seem to do it).